refactor(frontend): use `Link` instead of `navigate` where possible
