+++ web/frontend/src/components/SidebarItem.tsx
@@ -1,5 +1,6 @@
-import { Button, Link, Tooltip } from "@nextui-org/react"
+import { Button, Tooltip } from "@nextui-org/react"
 import { ReactNode, useState } from "react"
+import { Link } from "react-router-dom"
 import { FaCaretDown, FaCaretUp } from "react-icons/fa6"
 
 export type SidebarItemProps = {
@@ -15,18 +16,20 @@ export default function SidebarItem({ children, icon, isSidebarOpen, link, name 
 
   if(!isSidebarOpen) {
     return (
-      <Tooltip content="Services" size="sm" delay={1000}>
-        <Button as={Link} href={link} isIconOnly={true} variant="bordered" size="sm">
-          {icon}
-        </Button>
-      </Tooltip>
+      <Link to={link}>
+        <Tooltip content="Services" size="sm" delay={1000}>
+          <Button isIconOnly={true} variant="bordered" size="sm">
+            {icon}
+          </Button>
+        </Tooltip>
+      </Link>
     )
   }
 
   return (
     <div className="w-full flex flex-col gap-2">
       <div className="flex items-center">
-        <Link color="foreground" href={link} className="flex items-center gap-2">
+        <Link color="foreground" to={link} className="flex items-center gap-2">
           {icon}
           <span className="font-bold">{name}</span>
         </Link>
